map record class with i/o
