+++ app/src/main/java/com/scepticalphysiologist/dmaple/ui/Explorer.kt
@@ -1,0 +1,4 @@
+package com.scepticalphysiologist.dmaple.ui
+
+class Explorer {
+}
